Move MR3 timeline classes into Hive package
