+++ ql/src/java/org/apache/hadoop/hive/ql/exec/mr3/timeline/EntityIdentifier.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package com.datamonad.mr3.timeline;
+package org.apache.hadoop.hive.ql.exec.mr3.timeline;
 
 /**
  * The unique identifier for an entity
