+++ ql/src/java/org/apache/hadoop/hive/ql/exec/mr3/timeline/TimelineWriter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package com.datamonad.mr3.timeline;
+package org.apache.hadoop.hive.ql.exec.mr3.timeline;
 
 import java.io.IOException;
