+++ service/src/java/org/apache/hive/service/server/MR3TimelineService.java
@@ -21,17 +21,17 @@ package org.apache.hive.service.server;
 import java.io.IOException;
 import java.net.URL;
 
-import com.datamonad.mr3.timeline.AMProxyResource;
-import com.datamonad.mr3.timeline.ATSResource;
-import com.datamonad.mr3.timeline.MR3LiveStatusService;
-import com.datamonad.mr3.timeline.MR3TimelineIngestionService;
-import com.datamonad.mr3.timeline.LeveldbTimelineStore;
-import com.datamonad.mr3.timeline.MemoryTimelineStore;
-import com.datamonad.mr3.timeline.MR3LiveStatusServiceInterface;
-import com.datamonad.mr3.timeline.MR3TimelineIngestionServiceInterface;
-import com.datamonad.mr3.timeline.ServerResource;
-import com.datamonad.mr3.timeline.TimelineDataManager;
-import com.datamonad.mr3.timeline.TimelineStore;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.AMProxyResource;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.ATSResource;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.MR3LiveStatusService;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.MR3TimelineIngestionService;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.LeveldbTimelineStore;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.MemoryTimelineStore;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.MR3LiveStatusServiceInterface;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.MR3TimelineIngestionServiceInterface;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.ServerResource;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.TimelineDataManager;
+import org.apache.hadoop.hive.ql.exec.mr3.timeline.TimelineStore;
 import com.sun.jersey.api.core.ClassNamesResourceConfig;
 import com.sun.jersey.api.core.ResourceConfig;
 import com.sun.jersey.spi.container.servlet.ServletContainer;
